removed circular code dependencies. added lang tags to tree bodys and search box
