+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, AfterViewInit, ViewChild } from '@angular/core';
-import { GraphService } from './graph.service';
+import { DialogsService } from './dialogs.service';
 
 
 @Component({
@@ -12,18 +12,21 @@ export class AppComponent implements AfterViewInit {
 
 	@ViewChild('sidenav') public sidenav: any;
 
-	constructor(public graphService: GraphService) {
+	constructor(public dialogsService: DialogsService) {
 	}
 
-	ngAfterViewInit() {
-    // If they've never dismissed the declaimer
-    if (this.graphService.showDisclaimer) {
-      // Show the menu first
-			this.sidenav.open();
+  ngAfterViewInit() {
+    // We cannot immediately interact with the sidenav during init.
+    setTimeout(() => {
+      // If they've never dismissed the declaimer
+      if (this.dialogsService.showDisclaimer) {
+        // Show the menu first
+			  this.sidenav.open();
 
-			// Then show the disclaimer on top
-			this.graphService.openDialog(null, 'disclaimer');
-		}
+			  // Then show the disclaimer on top
+			  this.dialogsService.openDialog('disclaimer');
+		  }
+    }, 100);
 	}
 
   closeSidenav() {
