+++ src/app/graph.service.ts
@@ -12,19 +12,6 @@ import { TreeNode, IActionMapping } from 'angular-tree-component';
 import { GraphTab } from './GraphTab';
 import { GraphFilter } from './GraphFilter';
 
-import { MatDialog } from '@angular/material/dialog';
-import { AboutDialogComponent } from './dialogs/about-dialog.component';
-import { ChangeLogDialogComponent } from './dialogs/changelog-dialog.component';
-import { ContributeDialogComponent } from './dialogs/contribute-dialog.component';
-import { CreditsDialogComponent } from './dialogs/credits-dialog.component';
-import { DisclaimerDialogComponent } from './dialogs/disclaimer-dialog.component';
-import { DownloadDialogComponent } from './dialogs/download-dialog.component';
-import { HowToDialogComponent } from './dialogs/howto-dialog.component';
-import { ErrorsDialogComponent } from './dialogs/errors-dialog.component';
-import { PurchaseDialogComponent } from './dialogs/purchase-dialog.component';
-
-import { CookieService } from 'ngx-cookie-service';
-
 export interface ICategory {
     id: string;
     title: string;
@@ -67,14 +54,6 @@ export interface DAG {
 }
 // -- Dag Node --
 
-export class VisibleLink {
-    constructor(
-         public fromNode: TreeNode,
-         public link: Link) {
-         
-         }
-}
-
 @Injectable({ providedIn: 'root' })
 export class GraphService {
   private docGuids = {};
@@ -84,11 +63,8 @@ export class GraphService {
   private runningFilters = false;
   public updateSubject = new Rx.BehaviorSubject(0);
   public updateViewSubject = new Rx.BehaviorSubject(0);
-  private enableCookies = false;
 
   constructor(
-		public dialog: MatDialog,
-	  public cookies: CookieService,
     private messageService: MessageService,
     private http: HttpClient) {
       this.addTab("ISO");
@@ -444,36 +420,5 @@ export class GraphService {
           return true;
 
       return false;
-  }
-
-	public openDialog(dialogId: string) {
-		//sideNav.close();
-
-		var dialogType = null;
-    switch (dialogId) {
-			case 'about': dialogType = AboutDialogComponent; break;
-			case 'changelog': dialogType = ChangeLogDialogComponent; break;
-			case 'contribute': dialogType = ContributeDialogComponent; break;
-			case 'credits': dialogType = CreditsDialogComponent; break;
-			case 'disclaimer': dialogType = DisclaimerDialogComponent; break;
-			case 'download': dialogType = DownloadDialogComponent; break;
-			case 'howto': dialogType = HowToDialogComponent; break;
-      case 'purchase': dialogType = PurchaseDialogComponent; break;
-      case 'errors': dialogType = ErrorsDialogComponent; break;
-		}
-
-		if (dialogType) {
-			const dialogRef = this.dialog.open(dialogType);
-
-			dialogRef.afterClosed().subscribe(result => {
-				console.log(`Dialog result: ${result}`);
-        if (this.enableCookies && dialogId == 'disclaimer')
-					this.cookies.set('dismisseddisclaimer', 'y');
-			});
-		}
-  }
-
-  public get showDisclaimer(): boolean {
-    return !this.enableCookies || !this.cookies.get('dismisseddisclaimer');
   }
 }
